Only check writeback for `rd != 0`
RVFI states that `rvfi_rd_wdata` MUST be zero when `rd` is zero. However, CVA6
does not seem to follow this rule. Specifically, stores output the value
written to memory on that signal instead. But, this is not a correctness issue
since the destination register is zero.

To accomodate this implementation, this commit only checks that `rvfi_rd_wdata`
matches the specification if the destination register is non-zero. If it is
zero, the write will be ignored, so we don't need to check the value.